pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentic-workstation"
version = "1.0.4"
description = "Multi-agent planning and workflow orchestration system"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Sujith", email = "sujith.eag@gmail.com" }]

dependencies = [
    "click>=8.1.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",         # For ledger YAML files
    "jsonschema>=4.17.0",    # Schema validation
    "rich>=13.0.0",          # CLI formatting
    "rich-click>=1.7.0",     # Rich-styled CLI help
    "questionary>=2.0.0",    # Interactive prompts for TUI
    "structlog>=25.0.0",     # Structured logging
    "tomli_w>=1.0.0",        # TOML writing support
    "toml>=0.10.0",          # Fallback TOML support
    "tomli>=1.2.0",          # TOML reading for Python <3.11
    "platformdirs>=3.0.0",   # Cross-platform user dirs
    "pydantic>=2.0.0",       # Type-safe configuration models
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]
web = [
    "fastapi>=0.100.0",      # Web API framework
    "uvicorn>=0.23.0",       # ASGI server
]
auth = [
    "pyjwt>=2.8.0",          # JWT authentication
]
performance = [
    "cachetools>=5.3.0",     # Caching utilities
]
interactive = [
    "click-completion>=0.5.0", # Shell completion
]
external = [
    "requests>=2.31.0",      # HTTP client
]

[project.scripts]
agentic = "agentic_workflow.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_workflow"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""